webapp/backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
boto3==1.18.0
python-dotenv==0.19.0
gunicorn==20.1.0
werkzeug==2.0.3
flask-cognito
flask-awscognito